+++ Source/EbsdLib/Math/EbsdMatrixMath.cpp
@@ -52,7 +52,7 @@ EbsdMatrixMath::~EbsdMatrixMath() = default;
 // -----------------------------------------------------------------------------
 void EbsdMatrixMath::Multiply3x3with3x1(const double g1[3][3], const double g2[3], float outMat[3])
 {
-  outMat[0] = g1[0][0] * g2[0] + g1[0][1] * g2[1] + g1[0][2] * g2[2];
-  outMat[1] = g1[1][0] * g2[0] + g1[1][1] * g2[1] + g1[1][2] * g2[2];
-  outMat[2] = g1[2][0] * g2[0] + g1[2][1] * g2[1] + g1[2][2] * g2[2];
+  outMat[0] = static_cast<float>(g1[0][0] * g2[0] + g1[0][1] * g2[1] + g1[0][2] * g2[2]);
+  outMat[1] = static_cast<float>(g1[1][0] * g2[0] + g1[1][1] * g2[1] + g1[1][2] * g2[2]);
+  outMat[2] = static_cast<float>(g1[2][0] * g2[0] + g1[2][1] * g2[1] + g1[2][2] * g2[2]);
 }
